Actualizar pidController
